Load container architecture in system info
Make the Alpine Linux container architecture from the Alpine package
manager available in the system info. This information is useful not
only to create the deprecation repairs, but also as a general
system information since the container architecture may differ from
the host architecture, for example, when running a 64-bit host with
a 32-bit container.